@@ -1,8 +1,12 @@
 import { render, screen } from '@testing-library/react';
-import App from './App';
+import { Provider } from 'react-redux';
+import { BrowserRouter } from 'react-router-dom';
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
+import React from "react";
+import SamuraiJSApp from './App';
+
+test('renders Menu text', async () => {
+  render(<SamuraiJSApp />);
+  // const linkElement = screen.getByText(/proverka/i);
+  // expect(linkElement).toBeInTheDocument();
 });
